requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.2
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.5
django-countries==7.5.1
django-sites==0.11
executing==2.0.1
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.1
python-dotenv==1.0.0
pytils==0.4.1
redis==5.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.9.0
tzdata==2023.3
wcwidth==0.2.10
